+++ changelog.md
@@ -3,6 +3,7 @@
 #### Modifications actions
 
 - jcontent_by_categories : caché l'article en cours et article en cours indéterminé : essayer de le récupérer par getInput
+- sql : ajout pagination, setlimit accepte un nombre ou une requête sql
 
 ## 08/11/2025 - version 5.4.4